Refactor Megatron PP patches into parallelism package
